+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - Auto-process style, type tree and category index
 - Support loading JSON from a file if the string passed to the ``basic:json:``
   field is a file.
+- ``list:basic:integer:`` field
 
 Changed
 -------